+++ iOS/Flows/Channels/Cells/ConnectionCell.swift
@@ -1,0 +1,13 @@
+//
+//  ConnectionCell.swift
+//  Ours
+//
+//  Created by Benji Dodgson on 1/27/21.
+//  Copyright © 2021 Benjamin Dodgson. All rights reserved.
+//
+
+import Foundation
+
+class ConnectionCell: CollectionViewManagerCell {
+    
+}
