+++ iOS/Flows/Channels/ChannelsCollectionViewManager.swift
@@ -13,7 +13,7 @@ import Combine
 class ChannelsCollectionViewManager: CollectionViewManager<ChannelCell> {
 
     var didSelectReservation: ((Reservation) -> Void)? = nil
-    private var cancellables = Set<AnyCancellable>()
+    var cancellables = Set<AnyCancellable>()
 
     override func collectionView(_ collectionView: UICollectionView,
                                  layout collectionViewLayout: UICollectionViewLayout,
@@ -51,59 +51,5 @@ class ChannelsCollectionViewManager: CollectionViewManager<ChannelCell> {
                 return self.makeNonCurrentUserMenu(for: channel, at: indexPath)
             }
         })
-    }
-
-    private func makeCurrentUsertMenu(for channel: DisplayableChannel, at indexPath: IndexPath) -> UIMenu {
-
-        let neverMind = UIAction(title: "Never Mind", image: UIImage(systemName: "nosign")) { _ in }
-
-        let confirm = UIAction(title: "Confirm", image: UIImage(systemName: "trash"), attributes: .destructive) { action in
-
-            switch channel.channelType {
-            case .system(_):
-                self.delete(item: channel)
-            case .pending(_):
-                break 
-            case .channel(let tchChannel):
-                ChannelSupplier.shared.delete(channel: tchChannel)
-                    .mainSink().store(in: &self.cancellables)
-            }
-        }
-
-        let deleteMenu = UIMenu(title: "Delete", image: UIImage(systemName: "trash"), options: .destructive, children: [confirm, neverMind])
-
-        let open = UIAction(title: "Open", image: UIImage(systemName: "arrowshape.turn.up.right")) { _ in
-            self.select(indexPath: indexPath)
-        }
-
-        // Create and return a UIMenu with the share action
-        return UIMenu(title: "Options", children: [open, deleteMenu])
-    }
-
-    private func makeNonCurrentUserMenu(for channel: DisplayableChannel, at indexPath: IndexPath) -> UIMenu {
-
-        let neverMind = UIAction(title: "Never Mind", image: UIImage(systemName: "nosign")) { _ in }
-
-        let confirm = UIAction(title: "Confirm", image: UIImage(systemName: "clear"), attributes: .destructive) { action in
-
-            switch channel.channelType {
-            case .system(_):
-                self.delete(item: channel)
-            case .pending(_):
-                break 
-            case .channel(let tchChannel):
-                ChannelSupplier.shared.leave(channel: tchChannel)
-                    .mainSink().store(in: &self.cancellables)
-            }
-        }
-
-        let deleteMenu = UIMenu(title: "Leave", image: UIImage(systemName: "clear"), options: .destructive, children: [confirm, neverMind])
-
-        let open = UIAction(title: "Open", image: UIImage(systemName: "arrowshape.turn.up.right")) { _ in
-            self.select(indexPath: indexPath)
-        }
-
-        // Create and return a UIMenu with the share action
-        return UIMenu(title: "Options", children: [open, deleteMenu])
     }
 }
